Add CoT score function
